WIP krill/: Remove duplicate C64 and all C16 disk images
The .d71 and .d81 images for the C64 just duplicate the contents
of the .d64 images, so we can regenerate those if we need them.
We're not interested in the C16 for the purposes of this repo.

But these files compress _very_ well, so we don't mind leaving
them in the history.